feat: Add support for reading analog input from potentiometer
This commit adds the necessary code to read analog input from a potentiometer. It includes the creation of a new file `potentiometer_1.py` which contains the code for reading the analog input and displaying the value on a Tkinter GUI. The code uses the `pysmu` library for interacting with the ADAML1000 device.